backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
fastapi==0.115.0
uvicorn==0.32.0
qdrant-client==1.12.1
openai>=1.54.0
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
better-exceptions==0.3.2
sentence-transformers==3.1.1
transformers==4.45.2
psycopg2-binary>=2.9.0
sqlalchemy==2.0.35
alembic==1.13.3
markdown==3.7
email-validator==2.2.0
bcrypt==4.0.1
httpx>=0.27.0
